Refactor namespace from service.statement to reports.statement
- Move StatementModel and StatementReport to com.tvm.reportrendering.reports.statement package
- Update package declarations in both model and report classes
- Update all import statements in test files
- Fix API endpoint paths in integration tests (remove /api/ prefix)
- Fix error message expectations in controller integration tests
- All 48 tests passing after namespace refactor
