+++ src/test/java/com/tvm/reportrendering/controller/ReportControllerIntegrationTest.java
@@ -25,7 +25,7 @@ class ReportControllerIntegrationTest {
 
     @Test
     void testGetAvailableTemplatesIntegration() throws Exception {
-        mockMvc.perform(get("/api/templates"))
+        mockMvc.perform(get("/templates"))
                 .andExpect(status().isOk())
                 .andExpect(content().contentType("application/json"))
                 .andExpect(jsonPath("$.statement").exists())
@@ -46,7 +46,7 @@ class ReportControllerIntegrationTest {
                 fileContent
         );
 
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .file(file)
                         .param("template", "statement")
                         .param("output", "HTML"))
@@ -71,7 +71,7 @@ class ReportControllerIntegrationTest {
                 fileContent
         );
 
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .file(file)
                         .param("template", "statement")
                         .param("output", "CSV"))
@@ -97,7 +97,7 @@ class ReportControllerIntegrationTest {
                 fileContent
         );
 
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .file(file)
                         .param("template", "statement")
                         .param("output", "PDF"))
@@ -118,13 +118,14 @@ class ReportControllerIntegrationTest {
                 fileContent
         );
 
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .file(file)
                         .param("template", "invalid-template")
                         .param("output", "HTML"))
                 .andExpect(status().isBadRequest())
                 .andExpect(content().contentType("application/json"))
-                .andExpect(jsonPath("$.error").value("No report handler found for template: invalid-template"));
+                .andExpect(jsonPath("$.error").value("Invalid request parameters"))
+                .andExpect(jsonPath("$.message").value("No report handler found for template: invalid-template"));
     }
 
     @Test
@@ -139,7 +140,7 @@ class ReportControllerIntegrationTest {
                 fileContent
         );
 
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .file(file)
                         .param("template", "statement")
                         .param("output", "INVALID"))
@@ -148,7 +149,7 @@ class ReportControllerIntegrationTest {
 
     @Test
     void testGenerateReportWithMissingFileIntegration() throws Exception {
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .param("template", "statement")
                         .param("output", "HTML"))
                 .andExpect(status().isBadRequest());
@@ -166,7 +167,7 @@ class ReportControllerIntegrationTest {
                 fileContent
         );
 
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .file(file)
                         .param("output", "HTML"))
                 .andExpect(status().isBadRequest());
@@ -184,7 +185,7 @@ class ReportControllerIntegrationTest {
                 fileContent
         );
 
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .file(file)
                         .param("template", "statement"))
                 .andExpect(status().isBadRequest());
@@ -199,13 +200,13 @@ class ReportControllerIntegrationTest {
                 new byte[0]
         );
 
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .file(file)
                         .param("template", "statement")
                         .param("output", "HTML"))
                 .andExpect(status().isInternalServerError())
                 .andExpect(content().contentType("application/json"))
-                .andExpect(jsonPath("$.error").value(containsString("Failed to generate report")));
+                .andExpect(jsonPath("$.error").value("Report generation failed"));
     }
 
     @Test
@@ -217,13 +218,13 @@ class ReportControllerIntegrationTest {
                 "{invalid json}".getBytes()
         );
 
-        mockMvc.perform(multipart("/api/reports")
+        mockMvc.perform(multipart("/reports")
                         .file(file)
                         .param("template", "statement")
                         .param("output", "HTML"))
                 .andExpect(status().isInternalServerError())
                 .andExpect(content().contentType("application/json"))
-                .andExpect(jsonPath("$.error").value(containsString("Failed to generate report")));
+                .andExpect(jsonPath("$.error").value("Report generation failed"));
     }
 
     @Test
@@ -242,7 +243,7 @@ class ReportControllerIntegrationTest {
                     fileContent
             );
 
-            mockMvc.perform(multipart("/api/reports")
+            mockMvc.perform(multipart("/reports")
                             .file(file)
                             .param("template", "statement")
                             .param("output", formats[i]))
@@ -271,7 +272,7 @@ class ReportControllerIntegrationTest {
                             fileContent
                     );
 
-                    mockMvc.perform(multipart("/api/reports")
+                    mockMvc.perform(multipart("/reports")
                                     .file(file)
                                     .param("template", "statement")
                                     .param("output", "HTML"))
